+++ multichain-aggregator/multichain-aggregator-entity/src/dapps.rs
@@ -6,7 +6,7 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "dapps")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub chain_id: i32,
+    pub chain_id: i64,
     #[sea_orm(primary_key, auto_increment = false)]
     pub name: String,
     pub description: String,
