+++ multichain-aggregator/multichain-aggregator-entity/src/hashes.rs
@@ -13,7 +13,7 @@ pub struct Model {
     )]
     pub hash: Vec<u8>,
     #[sea_orm(primary_key, auto_increment = false)]
-    pub chain_id: i32,
+    pub chain_id: i64,
     pub hash_type: HashType,
     pub created_at: DateTime,
 }
